users 테이블에서 user 테이블로 경로 변경

+++ src/main/java/com/road_friends/rentalcar/mapper/APIUserMapper.java
@@ -10,6 +10,6 @@ import java.util.Optional;
 @Mapper
 public interface APIUserMapper {
   void save(UserDTO userDTO);
-  void insertUserRole(@Param("userId") int userId, @Param("roleId") int roleId);
+  void insertUserRole(@Param("userId") Long userId, @Param("roleId") int roleId);
 
 }
